databricks-nemweb-analyst-lab/app/requirements.txt
dash>=2.14.0
plotly>=5.18.0
pandas>=2.0.0
gunicorn>=21.0.0
databricks-sql-connector>=3.0.0
